requirements.txt
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
mypy-extensions==0.4.3
paramiko==2.10.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.5.1
pycparser==2.21
PyNaCl==1.5.0
python-nmap==0.7.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
